requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose
pydantic
opencv-python
deepface
pillow
pytesseract
python-dotenv